[fix] switch to int64

+++ backend/transaction/endpoint/histroy.go
@@ -9,8 +9,8 @@ import (
 )
 
 type match struct {
-	sellPrice    int
-	quantity     int
+	sellPrice    int64
+	quantity     int64
 	security     string
 	creationDate time.Time
 }
@@ -86,7 +86,7 @@ func OrderHistory(c *gin.Context, db *sql.DB) {
 		matchesJson[i] = gin.H{
 			"quantity": m.quantity,
 			"price":    m.sellPrice,
-			"created":  uint64(m.creationDate.Unix()),
+			"created":  m.creationDate.Unix(),
 			"security": m.security,
 		}
 	}
@@ -108,7 +108,7 @@ func AllHistory(c *gin.Context, db *sql.DB) {
 		matchesJson[i] = gin.H{
 			"quantity": m.quantity,
 			"price":    m.sellPrice,
-			"created":  uint64(m.creationDate.Unix()),
+			"created":  m.creationDate.Unix(),
 			"security": m.security,
 		}
 	}
